requirements.txt
cffi==1.15.0
click==8.0.3
ConcurrentLogHandler==0.9.1
Flask==2.0.2
Flask-Log-Request-ID==0.10.1
gunicorn==20.1.0
heif-image-plugin==0.3.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
piexif==1.1.3
pillow-avif-plugin==1.2.1
Pillow-SIMD==8.4.0.post0
pycparser==2.21
pydantic==1.8.2
pyheif==0.6.1
python-dotenv==0.19.2
typing-extensions==4.0.1
Werkzeug==2.0.2
